ver. 0.57: - в окне формирования продажи добавил возможность не оплачивать взрослый билет, если дети проходят на территорию города самостоятельно; - исправил формирование печатной формы билета, добавил qr-код; - исправил сохранение билетов в БД (добавил статус "бесплатный"); - увеличил шрифт в форме продаж и главной форме + скролл в главной форме (продажи); - изменил логику продления билетов многодетным; - исправил ошибку в закрытии смены; - исправил ошибку связанную с регистрацией позиций в кассовом чеке для билетов со скидкой; - добавил кнопки copy/paste фамилии при добавлении нового клиента; - отключил показ всплывающего окна с предложением открыть новую продажу; - исправил ошибки в открытии ранее сохраненной продажи; - добавил возможность изменения данных посетителя из формы продажи; - небольшие исправления кода --> PEP8; - заменил tableview на tablewidget в главной форме (исправил функции фильтрации); - изменил механизм сохранения информации в БД для отслеживания ошибок; - изменил запрос на получение актуального id, при сохранении новых данных; - исправил блокировку кнопок в окне продажи; - при оплате добавил выбор: печатать кассовый/банковский чек или нет.

+++ db/models/sale.py
@@ -27,7 +27,7 @@ class Sale(Base):
     )
     id_client = Column(Integer, ForeignKey('client.id'), comment='Id клиента')
     id_user = Column(Integer, ForeignKey('user.id'), comment='Id пользователя')
-    price = Column(SmallInteger, comment='Цена')
+    price = Column(Integer, comment='Цена')
     datetime = Column(DateTime, comment='Дата и время продажи')
     discount = Column(SmallInteger, comment='Скидка')
     status = Column(SmallInteger, comment='Статус продажи')
